correcting negative time values, iterations and interactions became long, instead of int.

+++ src/main/java/br/edu/cultural/simulation/CultureDisseminationSimulation.java
@@ -28,8 +28,8 @@ public abstract class CultureDisseminationSimulation implements Runnable {
 	// Statistics
 	protected long sim_start_ms;
 	protected long sim_finish_ms = -1;
-	protected int iterations = 0;
-	protected int interactions = 0;
+	protected long iterations = 0;
+	protected long interactions = 0;
 	
 	protected long stop_after_iterations = Long.MAX_VALUE;
 
@@ -291,15 +291,15 @@ public abstract class CultureDisseminationSimulation implements Runnable {
 		}
 	}
 
-	public int iterations() {
+	public long iterations() {
 		return iterations;
 	}
 	
-	public Integer interactions() {
+	public long interactions() {
 		return interactions;
 	}
 	
-	public int epoch() {
+	public long epoch() {
 		return this.iterations / this.nw.n_nodes;
 	}
